Add&move review multisignature tx to share
